requirements.txt
# Backend framework
fastapi==0.110.0
uvicorn[standard]==0.27.1

# Frontend
streamlit==1.31.1
requests==2.31.0

# OpenAI SDK
openai==1.12.0

# Vector store
faiss-cpu==1.7.4

# NLP utilities
nltk==3.8.1

# Document parsing
pypdf==4.0.1
python-docx==1.1.0

# Data handling
numpy==1.26.4

# Validation & schemas
pydantic==2.6.1

# Optional but useful
python-multipart==0.0.9
tqdm==4.66.1
python-dotenv
